wip: Finished writing new methods, untested
